requirements.txt
diffusers
transformers
accelerate
pillow
requests
hf_transfer
numpy
opencv-python
torch
huggingface-hub
